feat(adds separate mixin for conditional.py methods)
