+++ graphistry/plotter.py
@@ -6,10 +6,11 @@ from .feature_utils import FeatureMixin  # type: ignore
 from .dgl_utils import DGLGraphMixin  # type: ignore
 from .umap_utils import UMAPMixin  # type: ignore
 from .text_utils import SearchToGraphMixin  # type: ignore
+from compute.conditional import ConditionalMixin  # type: ignore
 
 mixins = ([
     CosmosMixin, NeptuneMixin, GremlinMixin, LayoutsMixin, 
-    SearchToGraphMixin, DGLGraphMixin, UMAPMixin, FeatureMixin,
+    SearchToGraphMixin, DGLGraphMixin, UMAPMixin, FeatureMixin, ConditionalMixin,
     ComputeMixin, PlotterBase, object
 ])
 
@@ -20,6 +21,7 @@ class Plotter(  # type: ignore
     def __init__(self, *args, **kwargs):
         PlotterBase.__init__(self, *args, **kwargs)
         ComputeMixin.__init__(self, *args, **kwargs)
+        ConditionalMixin.__init__(self, *args, **kwargs)
         FeatureMixin.__init__(self, *args, **kwargs)
         DGLGraphMixin.__init__(self, *args, **kwargs)
         UMAPMixin.__init__(self, *args, **kwargs)
